Refactor/112 (#119)
* remove: test 파일 삭제

* chore: import 오류 관련 코드 추가

* refactor: pages, component 코드 분리
