--- a.js
@@ -1,1 +1,0 @@
-a;
