+++ src/App.tsx
@@ -6,10 +6,9 @@ import Login from './pages/Login/index';
 import Navbar from './components/navbar';
 import SearchForm from 'components/navbar/searching/searchForm/SearchForm';
 import Main from 'components/main/Main';
-import SignUp from 'components/signUp/SignUp';
+import SignUp from 'pages/SignUp/index';
 import MyInfo from 'components/myInfo/MyInfo';
 function App() {
-
   useEffect(() => {
     console.log('rendered App');
   }, []);
